requirements.txt
confluent-kafka~=2.10.1
python-dotenv
sqlalchemy==1.4.54
psycopg2-binary==2.9.9
pandas
pyyaml
setuptools
databricks-connect==14.3.0
pyspark==3.4.1
py4j==0.10.9.7
apache-airflow==2.8.1
snowflake-sqlalchemy!=*
sqlalchemy-bigquery!=*
sqlalchemy-redshift!=*
sqlalchemy-spanner!=*
